+++ apps/tasks/services/google_tasks_service.py
@@ -19,18 +19,18 @@ def first_download_of_branch_information_task(branch_id):
     branch.is_detected = True
     branch.save()
     branch.name = ''.join(random.choice(letters) for i in range(10))
-    time.sleep(1)
+
     branch.address = ''.join(random.choice(letters) for i in range(10))
-    time.sleep(1)
+
     branch.phone_number = ''.join(random.choice(numbers) for i in range(10))
-    time.sleep(1)
+
     branch.email = ''.join(random.choice(letters) for i in range(10)) + '@' + ''.join(
         random.choice(letters) for i in range(5)) + '.ru'
-    time.sleep(1)
+
     branch.description = ''.join(random.choice(letters) for i in range(10))
-    time.sleep(1)
+
     branch.short_description = ''.join(random.choice(letters) for i in range(10))
-    time.sleep(1)
+
     branch.save()
     user_id = branch.company.owner_id
     uploading_branch_reviews_task.delay(user_id=user_id, branch_id=branch_id)
@@ -43,7 +43,7 @@ def uploading_branch_reviews_task(user_id, branch_id):
     connect = get_connect_by_id_branch_and_type(user=user, branch_id=branch_id, connect_type=2)
     if connect is None:
         raise Exception(f'Connect не найден для branch_id {branch_id}')
-    for i in range(100):
+    for i in range(1000):
         create_review_by_branch_id_service(
             user=user,
             branch_id=branch_id,
@@ -52,7 +52,7 @@ def uploading_branch_reviews_task(user_id, branch_id):
             rating=5,
             connect_id=connect.id
         )
-        time.sleep(1)
+
     return f'Успешно загружены 100 отзывов для branch_id {branch_id}'